refator baseFunction into a separate file
