+++ runtime/sema/basefunction.go
@@ -1,0 +1,54 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2020 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sema
+
+import (
+	"github.com/onflow/cadence/runtime/ast"
+	"github.com/onflow/cadence/runtime/common"
+)
+
+type baseFunction struct {
+	name           string
+	invokableType  InvokableType
+	argumentLabels []string
+}
+
+func (f baseFunction) ValueDeclarationType() Type {
+	return f.invokableType
+}
+
+func (baseFunction) ValueDeclarationKind() common.DeclarationKind {
+	return common.DeclarationKindFunction
+}
+
+func (baseFunction) ValueDeclarationPosition() ast.Position {
+	return ast.Position{}
+}
+
+func (baseFunction) ValueDeclarationIsConstant() bool {
+	return true
+}
+
+func (baseFunction) ValueDeclarationAvailable(_ ast.Location) bool {
+	return true
+}
+
+func (f baseFunction) ValueDeclarationArgumentLabels() []string {
+	return f.argumentLabels
+}
